+++ FROST-Server.MQTT.Moquette/src/main/java/de/fraunhofer/iosb/ilt/frostserver/mqtt/moquette/AuthWrapper.java
@@ -19,6 +19,7 @@ package de.fraunhofer.iosb.ilt.frostserver.mqtt.moquette;
 
 import static de.fraunhofer.iosb.ilt.frostserver.settings.CoreSettings.TAG_AUTH_ALLOW_ANON_READ;
 
+import de.fraunhofer.iosb.ilt.frostserver.service.InitResult;
 import de.fraunhofer.iosb.ilt.frostserver.settings.CoreSettings;
 import de.fraunhofer.iosb.ilt.frostserver.settings.Settings;
 import de.fraunhofer.iosb.ilt.frostserver.util.AuthProvider;
@@ -65,8 +66,9 @@ public class AuthWrapper implements IAuthenticator, IAuthorizatorPolicy {
         }
 
         @Override
-        public void init(CoreSettings coreSettings) {
+        public InitResult init(CoreSettings coreSettings) {
             // This dummy needs no initialisation.
+            return InitResult.INIT_OK;
         }
 
         @Override