+++ Plugins/BatchProcessing/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/batchprocessing/PluginBatchProcessing.java
@@ -20,6 +20,7 @@ package de.fraunhofer.iosb.ilt.frostserver.plugin.batchprocessing;
 import static de.fraunhofer.iosb.ilt.frostserver.util.HttpMethod.POST;
 
 import de.fraunhofer.iosb.ilt.frostserver.path.Version;
+import de.fraunhofer.iosb.ilt.frostserver.service.InitResult;
 import de.fraunhofer.iosb.ilt.frostserver.service.PluginRootDocument;
 import de.fraunhofer.iosb.ilt.frostserver.service.PluginService;
 import de.fraunhofer.iosb.ilt.frostserver.service.Service;
@@ -51,13 +52,14 @@ public class PluginBatchProcessing implements PluginService, PluginRootDocument,
     private boolean enabled;
 
     @Override
-    public void init(CoreSettings settings) {
+    public InitResult init(CoreSettings settings) {
         this.settings = settings;
         Settings pluginSettings = settings.getPluginSettings();
         enabled = pluginSettings.getBoolean(TAG_ENABLE_BATCH_PROCESSING, getClass());
         if (enabled) {
             settings.getPluginManager().registerPlugin(this);
         }
+        return InitResult.INIT_OK;
     }
 
     @Override