+++ Plugins/OData/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/odata/PluginOData.java
@@ -34,6 +34,7 @@ import de.fraunhofer.iosb.ilt.frostserver.path.Version;
 import de.fraunhofer.iosb.ilt.frostserver.plugin.odata.deserialize.JsonReaderOData;
 import de.fraunhofer.iosb.ilt.frostserver.plugin.odata.serialize.JsonWriterOdata40;
 import de.fraunhofer.iosb.ilt.frostserver.plugin.odata.serialize.JsonWriterOdata401;
+import de.fraunhofer.iosb.ilt.frostserver.service.InitResult;
 import de.fraunhofer.iosb.ilt.frostserver.service.PluginService;
 import de.fraunhofer.iosb.ilt.frostserver.service.RequestTypeUtils;
 import de.fraunhofer.iosb.ilt.frostserver.service.Service;
@@ -100,7 +101,7 @@ public class PluginOData implements PluginService, ConfigDefaults {
     private boolean enabled;
 
     @Override
-    public void init(CoreSettings settings) {
+    public InitResult init(CoreSettings settings) {
         this.settings = settings;
         Settings pluginSettings = settings.getPluginSettings();
         enabled = pluginSettings.getBoolean(TAG_ENABLE_ODATA, getClass());
@@ -108,7 +109,7 @@ public class PluginOData implements PluginService, ConfigDefaults {
             settings.getPluginManager().registerPlugin(this);
             new PluginResultFormatOData().init(settings);
         }
-
+        return InitResult.INIT_OK;
     }
 
     @Override