+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Implemented OData any() filters.
 * Added options for adding extra meta data to the landing page using the modelLoader plugin.
 * Implemented $filter for MQTT topics. Can be enabled using the `mqtt.allowFilter` setting.
+* Plugins can be loaded in any order.
 
 **Internal changes & Bugfixes**
 * Improved generated queries when fetching entities over a one-to-many relation.